requirements.txt
torch
numpy
opencv-python
matplotlib
Pillow
moviepy